Refactor package namespace to include `web` as a subpackage for extensibility.
